+++ plugin/src/main/kotlin/tech/skot/Versions.kt
@@ -4,7 +4,7 @@ import kotlin.Int
 import kotlin.String
 
 public object Versions {
-  public const val skot: String = "1.2.29-ua"
+  public const val skot: String = "1.2.30-ua"
 
   public const val group: String = "com.github.useradgents.skot"